Add files via upload
This is the model of my LIFI device, this transmit data in the speed of light, without any, wifi or bluetooth connection, you can see, there is no connection between both the circuits, transmitter or receiver, neither magnetic field, it is totally sending and receiving data through light, which is faster than, wifi, signal, bluetooth signal, and magnetic field. This experiment was created in my bachelors or electrical.